requirements/requirements-api-ml.txt
fastapi==1.10.2
openai==0.27.0
paddlepaddle==2.4.2
paddleocr==2.6.1.3
python-dotenv==1.0.0
python-multipart==0.0.5
sentencepiece==0.1.99
torch==2.0.1
transformers==4.29.2
uvicorn==0.20.0

